requirements.txt
fastapi==0.120.0
uvicorn==0.38.0
pydantic==2.12.3
openai==1.58.1
google-api-python-client==2.150.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
websockets==15.0.1
python-multipart==0.0.20
python-dotenv==1.1.1
httpx==0.28.1
asyncio-mqtt==0.16.2
aiofiles==24.1.0
pydub==0.25.1
speechrecognition==3.10.4
pyttsx3==2.90
